Refactor history agent into modular package
