+++ src/history_agent/__init__.py
@@ -1,0 +1,6 @@
+"""History research agent package."""
+
+from .history_graph import research_agent
+
+__all__ = ["research_agent"]
+
